Set up support for adding event listeners
both to the created elment as a load event

and also as an available function for specifying on other elements
would like to make it so maybe
events can be applied using promises?
so it would be var.addEvent().load()
that way we could control when the script was loaded? needs more research